feat: Migrations data skips silver layer, directly to Gold (#597)
If applied, this PR changes the 
destination of migrations data from 
silver to gold. 
The reason being that the merge keys 
from migrations do not match with 
what Submitted expects in silver, 
and it makes better architecture-design 
sense to have the different 
orchestration_types meet in gold rather 
than silver.